Split OpRegistry into model format-specific code and wrapper
In preparation for adding direct support for .onnx files, split the operator
registry into a type that is specific to the .rten FlatBuffers format and a
format-independent wrapper.
