+++ src/lib.rs
@@ -142,7 +142,7 @@ pub use buffer_pool::{BufferPool, ExtractBuffer, PoolRef};
 pub use graph::{Dimension, NodeId, RunError, RunOptions};
 pub use model::{Model, ModelLoadError, ModelOptions, NodeInfo};
 pub use model_metadata::ModelMetadata;
-pub use op_registry::{OpRegistry, ReadOp, ReadOpError};
+pub use op_registry::{OpRegistry, ReadOpError};
 pub use ops::{FloatOperators, Operators};
 pub use threading::{ThreadPool, thread_pool};
 pub use timing::TimingSort;
